requirements.txt
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.2.2
fonttools==4.56.0
importlib_resources==6.5.2
iniconfig==2.0.0
kiwisolver==1.4.7
line_profiler==4.2.0
llvmlite==0.43.0
matplotlib==3.9.4
numba==0.60.0
numpy==2.0.2
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.13.1
six==1.17.0
tomli==2.2.1
zipp==3.21.0
